content/html/content/test/test_bug460568.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=460568
-->
<head>
  <title>Test for Bug 460568</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=460568">Mozilla Bug 460568</a>
<p id="display"><a href="" id="anchor">a[href]</a></p>
<div id="editor">
  <a href="" id="anchorInEditor">a[href] in editor</a>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 460568 **/

function runTest()
{
  var editor = document.getElementById("editor");
  var anchor = document.getElementById("anchor");
  var anchorInEditor = document.getElementById("anchorInEditor");

  var focused;
  anchorInEditor.onfocus = function() { focused = true; };

  focused = false;
  anchor.focus();
  editor.setAttribute("contenteditable", "true");
  anchorInEditor.focus();
  is(focused, false, "focus moved to element in contenteditable=true");

  focused = false;
  anchor.focus();
  editor.setAttribute("contenteditable", "false");
  anchorInEditor.focus();
  is(focused, true, "focus didn't move to element in contenteditable=false");

  focused = false;
  anchor.focus();
  editor.setAttribute("contenteditable", "true");
  anchorInEditor.focus();
  is(focused, false, "focus moved to element in contenteditable=true");

  focused = false;
  anchor.focus();
  editor.removeAttribute("contenteditable");
  anchorInEditor.focus();
  is(focused, true, "focus didn't move to element in contenteditable removed element");

  focused = false;
  anchor.focus();
  editor.contentEditable = true;
  anchorInEditor.focus();
  is(focused, false, "focus moved to element in contenteditable=true by property");

  focused = false;
  anchor.focus();
  editor.contentEditable = false;
  anchorInEditor.focus();
  is(focused, true, "focus didn't move to element in contenteditable=false by property");
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
